Dovecot restart cronjob
Adding a cronjob that each five minutes check if dovecot is runing, and
if not, will try to restart dovecot. This is due to the Time Moving
Backward "bug", which will make dovecot to kill itself.

Though there are some more complex cronscripts at
http://wiki.dovecot.org/TimeMovedBackwards, I felt it wasn't needed to
fork an new shell just to check, but merely using the init script status
and some shell magic.